scripts/requirements.txt
lyricsgenius==3.0.1
python-dotenv==1.0.1
